requirements.txt
flask==3.0.2
python-aiml==0.9.3
requests==2.32.3
